requirements.txt
langchain-openai
langchain
python-dotenv
langchain-community
langchain_core
streamlit
langchain-groq
ipykernel
pypdf
bs4
arxiv
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss_cpu
duckdb
pandas
openai
langchain-groq
tf-keras